Timer frequency (1000 HZ)
I know my own workloads & needs... dot dot dot

intentionally sacrificing maximum theoretical
throughput is a tradeoff: better to have the
decrease in latencies for multitasking. this
is somewhat expensive, due to context switching,
and other assorted "overhead thingers" [tm].

expecting a slight decrease in energy efficiency,
along with a proportional (trivial) increase in
CPU temperatures. I could measure this if I was
concerned, but this kernel isn't for android.

after further consideration, "race to idle" was
something which had mixed results in testing in
past android development. depends on workload(s)

it may or may not be MORE energy efficient. this
change is for the latencies though. whatever.

through-put shmou-put (@spelling?)

CONFIG_HZ=1000

I was not able to make sense of the following
kconfig change during a 30 minute code-dive:

CONFIG_HZ_100 --> CONFIG_HZ_1000_NODEF

Conclusion: 1000_NODEF is a benign/inert artifact
which seems to only exist for ck-style kconfigs:
unimplemented feature is my best guess for this?

renamed:

[old] 2k18M5.0940_AntiTickless.BC601756
[new] 2k18M5.1155_AntiTickless.D27FF25B