ZSWAP: Compressed cache for swap pages [@647E5141]
CONFIG_ZPOOL (API for zswap, etc.)

:[CONFIG_Z3FOLD]:
A special purpose allocator for storing compressed pages.
It is designed to store up to three compressed pages per
physical page. It is a ZBUD derivative so the simplicity
and determinism are still there.

:[CONFIG_FRONTSWAP]:
Frontswap is so named because it can be thought of as the
opposite of a "backing" store for a swap device. The data
is stored into "transcendent memory", memory that is not
directly accessible or addressable by the kernel and is of
unknown and possibly time-varying size. When space in
transcendent memory is available, a significant swap I/O
reduction may be achieved. When none is available, all
frontswap calls are reduced to a single pointer-
compare-against-NULL resulting in a negligible performance
hit and swap data is stored as normal on the matching swap
device.

:[CONFIG_ZSWAP]:
A lightweight compressed cache for swap pages. It takes
pages that are in the process of being swapped out and
attempts to compress them into a dynamically allocated
RAM-based memory pool. This can result in a significant
I/O reduction on swap device and, in the case where
decompressing from RAM is faster that swap device reads,
can also improve workload performance.

--------- ZRAM obsoleted ---------

WARNING: LRU inversion if ZRAM is mixed with "real" swap.

removed:

CONFIG_ZRAM
CONFIG_ZSMALLOC